Refactor: Organized services, schema and index
